requirements.txt
Django==3.0.8
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
dj-database-url==0.5.0
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pip==20.1.1
psycopg2==2.8.5
pylint==2.5.3
pytz==2020.1
setuptools==49.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1